pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["Ryosuke Yamamoto <ryosuke-yamamoto@fintec.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.1"
pandas = "^1.5.1"
aws = "^0.2.5"
awscli = "^1.27.59"
python-dotenv = "^0.21.1"
awswrangler = "^2.19.0"
boto3 = "^1.26.59"
isort = "^5.12.0"
networkx = "^3.0"
notebook = "^6.5.3"
tqdm = "^4.65.0"
matplotlib = "^3.7.1"
streamlit = "^1.20.0"
flask = "^2.2.3"
praw = "^7.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
